ECMP over multiple wg devices for an allowed-ip
This commit makes amesh create wg devices 'per-'endpoint peers.
If multiple peers have the same allowed-ips (prefixes), the
prefixes are used as an ECMP route.

Example:

% sudo ip netns exec netns2 ip r s
172.16.0.0/24 dev netns2-b proto kernel scope link src 172.16.0.2
172.16.1.0/24
	nexthop dev wg-rJTjofx5uE weight 1
	nexthop dev wg-6t5DLJtUjU weight 1

in this case, the two peers with pubkey rJTjo~ and 6t5DL~ are
installed on the different wg devices, and they have the same
allowed-ip 172.16.1.0/24. As a result, 172.16.1.0/24 becomes an
ECMP route via the two wg devices.

ref https://lists.zx2c4.com/pipermail/wireguard/2018-May/002804.html